Replace woodsmanship with woodworking
Looking through several dictionaries, it does not seem like
"woodsmanship" is not broadly accepted as a word. It seems to be a
neologism connected with the word "woodsman".

I also think "woodworking" is more inclusive than "woodsmanship".
Regardless of the etymology, "woodsman" is not a gender neutral word.
And I certainly know more technology folk who work **with** wood then
who work **in** the woods.